x_watch_monitor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "x-watch-monitor"
version = "0.1.0"
description = "Periodic X user monitoring with Grok analysis and Discord notifications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.2,<7",
  "python-dotenv>=1.0.1,<2",
  "requests>=2.32.3,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5,<9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--basetemp=pytest_tmp -p no:cacheprovider"
